fix: resolve dead code and implement alert sharing functionality
